+++ src/transform/layout_inference.cc
@@ -32,7 +32,6 @@
 #include "common/loop_fusion_utils.h"
 #include "common/pipeline_utils.h"
 #include "common/union_find.h"
-#include "layout_reducer.h"
 #include "parallel_loop_layout_validator.h"
 #include "span_utils.h"
 #include "tir/transforms/ir_utils.h"
